Bug fix



See merge request !86

+++ src/main/java/com/hoolink/manage/base/service/impl/UserServiceImpl.java
@@ -2,29 +2,7 @@ package com.hoolink.manage.base.service.impl;
 
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
-import com.hoolink.manage.base.bo.DeptPositionBO;
-import com.hoolink.manage.base.bo.DeptSecurityBO;
-import com.hoolink.manage.base.bo.DeptTreeBO;
-import com.hoolink.manage.base.bo.DictInfoBO;
-import com.hoolink.manage.base.bo.DictParamBO;
-import com.hoolink.manage.base.bo.LoginParamBO;
-import com.hoolink.manage.base.bo.LoginResultBO;
-import com.hoolink.manage.base.bo.ManageRoleBO;
-import com.hoolink.manage.base.bo.ManagerUserInfoBO;
-import com.hoolink.manage.base.bo.ManagerUserInfoParamBO;
-import com.hoolink.manage.base.bo.ManagerUserPageParamBO;
-import com.hoolink.manage.base.bo.ManagerUserParamBO;
-import com.hoolink.manage.base.bo.MiddleUserDepartmentBO;
-import com.hoolink.manage.base.bo.MiddleUserDeptWithMoreBO;
-import com.hoolink.manage.base.bo.PersonalInfoBO;
-import com.hoolink.manage.base.bo.PhoneParamBO;
-import com.hoolink.manage.base.bo.RoleMenuPermissionBO;
-import com.hoolink.manage.base.bo.UpdatePasswdParamBO;
-import com.hoolink.manage.base.bo.UserDeptBO;
-import com.hoolink.manage.base.bo.UserDeptPairBO;
-import com.hoolink.manage.base.bo.UserDeptPairParamBO;
-import com.hoolink.manage.base.bo.UserInfoBO;
-import com.hoolink.manage.base.bo.UserSecurityBO;
+import com.hoolink.manage.base.bo.*;
 import com.hoolink.manage.base.constant.Constant;
 import com.hoolink.manage.base.consumer.ability.AbilityClient;
 import com.hoolink.manage.base.consumer.edm.EdmClient;
@@ -564,16 +542,15 @@ public class UserServiceImpl implements UserService {
 		if(CollectionUtils.isEmpty(roleList)) {
 			return new PageInfo<ManagerUserBO>();
 		}
-		
+        PageHelper.startPage(userPageParamBO.getPageNo(), userPageParamBO.getPageSize());
+
         UserExample example = buildUserCriteria(userPageParamBO);
-        PageInfo<User> userPageInfo = PageHelper
-                .startPage(userPageParamBO.getPageNo(), userPageParamBO.getPageSize())
-                .doSelectPageInfo(() -> userMapper.selectByExample(example));
-        List<User> userList = userPageInfo.getList();
+        List<User> userList = userMapper.selectByExample(example);
+        PageInfo pageInfo = new PageInfo<>(userList);
 
         //组装用户数据
         List<ManagerUserBO> userBoList = buildUserBOList(userList);
-        PageInfo<ManagerUserBO> userBOPageInfo = CopyPropertiesUtil.copyPageInfo(userPageInfo, ManagerUserBO.class);
+        PageInfo<ManagerUserBO> userBOPageInfo = CopyPropertiesUtil.copyPageInfo(pageInfo, ManagerUserBO.class);
         userBOPageInfo.setList(userBoList);
         return userBOPageInfo;
     }
@@ -1178,6 +1155,21 @@ public class UserServiceImpl implements UserService {
         Long userId = getCurrentUserId();
         User user = buildUserToUpdate(userId);
         user.setPasswd(MD5Util.MD5(updatePasswdParam.getPasswd()));
+        userMapper.updateByPrimaryKeySelective(user);
+        sessionService.deleteRedisUser(userId);
+    }
+
+    @Override
+    @Transactional(rollbackFor = Exception.class)
+    public void updateMobilePassword(MobileUpdateParamBO mobileUpdateParamBO) {
+        Long userId = getCurrentUserId();
+        User user=userMapper.selectByPrimaryKey(userId);
+        if(!Objects.equals(MD5Util.MD5(mobileUpdateParamBO.getOldPassword()),user.getPasswd())){
+            throw new BusinessException(HoolinkExceptionMassageEnum.PASSWORD_ERROR);
+        }
+        user.setUpdator(ContextUtil.getManageCurrentUser().getUserId());
+        user.setUpdated(System.currentTimeMillis());
+        user.setPasswd(MD5Util.MD5(mobileUpdateParamBO.getNewPassword()));
         userMapper.updateByPrimaryKeySelective(user);
         sessionService.deleteRedisUser(userId);
     }
